+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Kill running process and its children recursively (gorarakelyan)
 - Keep executed processes for monitoring and management (gorarakelyan)
 - Add container command handler to exec commands on the host (gorarakelyan)
 - Refactor Text artifact to store sentences using protobuf and aimrecords (jamesj-jiao)